+++ app/web/js/main.js
@@ -14,6 +14,8 @@ $(function() {
             chartPadding: {
                 top: 0,
                 bottom: 0,
+                right: 0,
+                left: 0
             },
             distributeSeries: true,
             height: "290px",
