+++ run.py
@@ -1,0 +1,3 @@
+from app import app, db_utils
+
+app.run()
